requirements.txt
selenium==4.22.0
requests==2.32.3
deep_translator==1.11.4
pyotp==2.9.0
python-dateutil~=2.8.2
python-dotenv~=1.0.0
